+++ src/main/java/org/eclipse/tractusx/selfdescriptionfactory/service/converter/fcformat/ServiceOfferingSDConverter.java
@@ -1,6 +1,6 @@
 /********************************************************************************
- * Copyright (c) 2022,2023 T-Systems International GmbH
- * Copyright (c) 2022,2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022,2025 T-Systems International GmbH
+ * Copyright (c) 2022,2025 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -20,8 +20,10 @@
 
 package org.eclipse.tractusx.selfdescriptionfactory.service.converter.fcformat;
 
+import com.danubetech.verifiablecredentials.CredentialSubject;
+import com.danubetech.verifiablecredentials.VerifiableCredential;
 import lombok.RequiredArgsConstructor;
-import org.eclipse.tractusx.selfdescriptionfactory.SDFactory;
+import org.eclipse.tractusx.selfdescriptionfactory.SelfDescription;
 import org.eclipse.tractusx.selfdescriptionfactory.Utils;
 import org.eclipse.tractusx.selfdescriptionfactory.model.vrel3.ServiceOfferingSchema;
 import org.eclipse.tractusx.selfdescriptionfactory.service.converter.TermsAndConditionsHelper;
@@ -30,7 +32,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
-import java.util.List;
+import java.net.URI;
+import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -38,37 +41,43 @@ import java.util.function.Function;
 @Component
 @RequiredArgsConstructor
 @Profile("fc-ctx")
-public class ServiceOfferingSDConverter implements Converter<ServiceOfferingSchema, SDFactory.SelfDescription> {
+public class ServiceOfferingSDConverter implements Converter<ServiceOfferingSchema, SelfDescription> {
 
     private final CustodianWallet custodianWallet;
     private final TermsAndConditionsHelper termsAndConditionsHelper;
 
     @Override
-    public SDFactory.SelfDescription convert(ServiceOfferingSchema serviceOfferingSchema) {
-        var serviceOfferingSD = new SDFactory.SelfDescription(List.of(), serviceOfferingSchema.getHolder(), serviceOfferingSchema.getIssuer(), serviceOfferingSchema.getExternalId(), null);
-        serviceOfferingSD.put("@context", Map.of(
+    public SelfDescription convert(ServiceOfferingSchema serviceOfferingSchema) {
+
+
+        var serviceOfferingSD = new SelfDescription(serviceOfferingSchema.getExternalId());
+
+        var serviceOfferingVc = new LinkedHashMap<String, Object>();
+
+
+        serviceOfferingVc.put("@context", Map.of(
                         "gx", "https://w3id.org/gaia-x/gax-trust-framework#",
                         "gax-core", "https://w3id.org/gaia-x/core#",
                         "ctxsd", "https://w3id.org/catena-x/core#",
                         "xsd", "http://www.w3.org/2001/XMLSchema#"
                 )
         );
-        serviceOfferingSD.put("@id", custodianWallet.getWalletData(serviceOfferingSchema.getHolder()).get("did"));
-        serviceOfferingSD.put("@type", "gx:ServiceOffering");
-        serviceOfferingSD.put("ctxsd:connector-url", "https://connector-placeholder.net");
+        serviceOfferingVc.put("@id", custodianWallet.getWalletData(serviceOfferingSchema.getHolder()).get("did"));
+        serviceOfferingVc.put("@type", "gx:ServiceOffering");
+        serviceOfferingVc.put("ctxsd:connector-url", "https://connector-placeholder.net");
         //WARNING! In Trust Framework specs it is called 'providedBy'
         //but in FC it is referred as 'gax-core:offeredBy'
-        serviceOfferingSD.put("gax-core:offeredBy", Map.of(
+        serviceOfferingVc.put("gax-core:offeredBy", Map.of(
                 "@id", serviceOfferingSchema.getProvidedBy())
         );
-        serviceOfferingSD.put("gx:dataAccountExport", Map.of(
+        serviceOfferingVc.put("gx:dataAccountExport", Map.of(
                 "gx:requestType", "email",
                 "gx:accessType", "digital",
                 "gx:formatType", "json")
         );
         var setter = new Object() {
             Consumer<Object> set(String fieldName) {
-                return any -> serviceOfferingSD.put(fieldName, any);
+                return any -> serviceOfferingVc.put(fieldName, any);
             }
         };
         Utils.getNonEmptyListFromCommaSeparated(serviceOfferingSchema.getAggregationOf(), Utils::uriFromStr)
@@ -101,6 +110,13 @@ public class ServiceOfferingSDConverter implements Converter<ServiceOfferingSche
                         ).map(l -> l.size() == 1 ? l.iterator().next() : l)
                         .orElse("policy_placeholder")
         );
+
+        var legalParticipantSD = VerifiableCredential.builder()
+                .issuer(URI.create(serviceOfferingSchema.getIssuer()))
+                .credentialSubject(CredentialSubject.fromMap(serviceOfferingVc))
+                .build();
+        serviceOfferingSD.getVerifiableCredentialList().add(legalParticipantSD);
+
         return serviceOfferingSD;
     }
 }